OPC Candidates
We have engineering work at every layer of the stack, from REST APIs to distributed systems to file systems, and we are looking for teammates who are interested in being part of a team that innovates from top to bottom. We believe that the only way to succeed is to own every part of the problem, and so we are creating a team that controls its own destiny. Our team is located in the San Francisco Bay Area and Seattle.

Qualifications
Candidates should be passionate about distributed systems, cloud computing, and container management. Candidates will need to possess deep CS expertise, as well as excel at written and verbal communication.
•	BS in Computer Science; level commensurate with years of experience.
•	Demonstrable experience with Java, C++, or Go.
•	Experience with distributed systems: familiarity with CAP theorem, consensus, replication, etc.
•	Background in building, testing, and debugging robust production services.
•	Strong troubleshooting and performance tuning skills.